Updated API, Dashboard Map Visualization
The map report on the web dashboard is now zoomable to state level.
Click state to zoom in, click state again to zoom out.

Modified API parameter logic for start date and end date to be
inclusive instead of inclusive on the upper and lower bounds of the
date range.