Implemented new features
Dashboard:

* You can now select granularity, daily, weekly, and monthly.
* Changed time series to report on periodic new user count.
* Changed the time series bar chart visualization to use area chart.

Scheduler:

* Implemented group tracking. The scheduler now allows you to import
groups to track on by tag and location. Once a group has been imported
the first time, it will always be tracked regardless of if the tags
change. Previously a group could lose all of its members in one day if
a tracked tag was removed from the group.